SVN: merge from issue-cl-1825 branch to trunk-3-4 (int2long - number of records precision)
git-svn-id: https://svn.javlin.eu/svn/cloveretl.opensource/trunk-3-4@13683 a09ad3ba-1a0f-0410-b1b9-c67202f10d70

+++ cloveretl.component/src/org/jetel/component/SimpleDataGenerator.java
@@ -130,7 +130,7 @@ public class SimpleDataGenerator extends DataGenerator {
 	 * @param pattern
 	 * @param recordsNumber
 	 */
-	public SimpleDataGenerator(String id, String pattern, int recordsNumber) {
+	public SimpleDataGenerator(String id, String pattern, long recordsNumber) {
 		super(id);
 		this.pattern = pattern;
 		this.recordsNumber = recordsNumber;
@@ -177,7 +177,7 @@ public class SimpleDataGenerator extends DataGenerator {
 	public Result execute() throws Exception {
 		DataRecord record = DataRecordFactory.newRecord(getOutputPort(WRITE_TO_PORT).getMetadata());
 		record.init();
-		for (int i=0;i<recordsNumber && runIt;i++){
+		for (long i=0;i<recordsNumber && runIt;i++){
 			record = recordGenerator.getNext(record);
 			autoFilling.setAutoFillingFields(record);
 			writeRecordBroadcast(record);
